requirements.txt
streamlit==1.38.0
pandas==2.2.2
numpy==1.26.4
tensorflow-intel==2.16.1
scikit-learn==1.5.1
xgboost==2.1.1
plotly==5.24.1
joblib==1.4.2
openpyxl==3.1.5
pyarrow==17.0.0
